Finished Recreating the Outskirts Level Layout + Tweaked Player Ability
I finished recreating the base layout for the Outskirts level using tile maps (Thank you, Yue Sheng for that inspiration). However, of course objects and their placement will need to be more refined and tweaked for optimal gameplay. Lastly, I tweaked some code for the PlayerAbilitiesScript to allow the player to break breakable objects when dashing (Credit to Yue Sheng for the code).