Added all of Yue Sheng's Override changes to the Player prefab + More
As said, all of Yue Sheng's Override changes were added to the Player Prefab, so that the player can work the same across all scenes. Additionally, the SFX Override changes I made were added to the player prefab. Furthermore, I changed the player's size back to 10 and made that as a override change to make the player consistent across other scenes. Lastly, due to the change in player size, the positions of certain obstacles in the Outskirts Level were changed to better gameplay.