Due to Player Feedback Many Things Were Fixed
The Walking SFX was changed to Jamari's version. The Hit box of the Player was fixed and should hopefully make Gameplay better. The placement of some things were changed to better Gameplay. Lastly, the Text for the Tutorial was changed to hopefully help Players better understand the Mechanics and their abilities.