Reworked the Checkpoint System
Remade the Checkpoint System so that multiple or probably infinite amounts of Checkpoints can be added to a given level or scene. Checkpoints work by just simply adding multiple versions of the Checkpoint prefab and placing them at whichever location is desired. Lastly, on the "GameManager" Object in a given scene make  sure the starting "lastCheckpointPosition" is the position of where you want the player to start at the level.