Added Yue Sheng's Dash Glow as an Override + Changed one obstacle + More
Yue Sheng's New Dash Glow was made as an Override for the Player Prefab, that way the player can have the Dash Glow across all scenes. One obstacle was changed to make it a little easier and allow the player an opportunity to better utilize the diagonal dash and glide abilities.