Added the new Player Prefab to all levels + More Changes
As said, the new Player Prefab was added to all Levels. The stomp was changed to match the new movement. It is basically like jump, but instead you go vertically down. Fixed an issue that allowed the player to break breakable objects by simply just touching them after the stomp move was performed a long time before. Lastly, the Trail Renderer effect that was used for the Dash was added as an effect for the stomp move as well. The Trail Renderer effect may be used for the Glide in a future commit/update.