The "V" Key was added as an option to perform the Glide, instead of "K"
To compromise for certain users that would've liked to have a easier way to use the glide move I added the "V" key as an option because it is closer to the space bar, which is also jump. This should hopefully help because there is no other option that can make the glide activate using one key.