Finished the Forest Level Layout + More
As said, the Forest Level Layout is done. Additionally, some dash values for the Player was changed to make the dash less Overpowered and better for gameplay, but this change will be monitored if it does need further changes in the future. Lastly, the colliders on Checkpoints were changed to allow the Player to easily activate them when they pass by a checkpoint.