Fixed Animation of the Stomp + added a background to Outskirts + more
As said the animation of the Stomp was fixed. Specifically, the animation was being played backwards, which was my bad. Lastly, I added the Outskirts background to the Outskirts level. It's no parallax scrolling, but I attached the background to the camera, that way it could give an effect of the background following the player. This is only a temporary thing until we get around to making the backgrounds look better with parallax scrolling. Lastly, a slope was fixed at the beginning of the Outskirts to make a jump look smoother and easier for most players.