Added a Reputation System + Choice Menu Panel UI
A Reputation System was added. This allows the game to keep track of the Player's choices when Destroying or Leaving Dragon Eggs alone, which will determine what Ending Cutscene the Player will get. When the Player Destroys a Dragon Egg, their Reputation with the king will go up by one point. If they leave it alone, Reputation with the Dragons will go up by one point. The Choice Menu Panel is something else that was added that pops up in-game when the Player interacts with a Dragon Egg. For now, the Player can only make two choices: either Destroy or Leave The Dragon Egg Alone, which have designated buttons for the player to press with their mouse. More may be added to the Reputation System and the Choice Panel UI, but for now it works very well with its intended purpose.