Fixed Collison issues with levels + Added a tutorial for Stomp move
Fixed an issue that made collisions with breakable obstacles unusable in itch.io builds. A tutorial instruction for the Stomp Move was added to the Forest level. Lastly, animations for the Glide Move was added and fixed an issue that allowed the Glide SFX to be played on the ground even when not actually performing the Glide Move, but pressing the key for it.